refactor(front): centralize query layer (COS-157)
Move QUERY_KEYS out of the spendings module into a shared
front/src/lib/query/keys.ts (@lib/query/keys) so statistics,
exceptionals and categories no longer depend on spendings. Fold the
inline "regularMonthlyAverage" and "statistics" keys into it.

Dissolve QUERY_OPTIONS into the QueryClient defaultOptions.queries
(refetchOnWindowFocus, staleTime 1h) and enable throwOnError by
default; hooks keep only their own overrides (retry, enabled, gcTime,
placeholderData). Drop the manual `if (error) throw error` guards from
the components now that queries throw to the boundary on their own.

Search, label suggestions, category stats and spending years opt out
of throwOnError locally: their UI degrades gracefully and must not
trip the error screen. Expired sessions are unaffected — the 401 is
absorbed in useRequestHelper before it ever reaches React Query
(COS-26).
